% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASF_simulation.R
\name{ASF_simulation}
\alias{ASF_simulation}
\title{Simulation of ASF transmission in New Zealand feral pigs}
\usage{
ASF_simulation(
  total_iter,
  ncore,
  simyear,
  pop,
  scale,
  island,
  introloc,
  control_index,
  foi
)
}
\arguments{
\item{total_iter}{Number of total iterations. Recommend either 100 or 200.}

\item{ncore}{Number of CPU threads available for parallel computation.}

\item{simyear}{Simulation period in year. Input value should be an integer between one and five.}

\item{pop}{Total population size of feral pig. Input value should be either;
\itemize{
\item 0 (= small),
\item 1 (= intermediate),
\item 2 (= large).
}}

\item{scale}{Average habitat size of feral pigs. Input value should be either;
\itemize{
\item 0 (= 4km2),
\item 1 (= 10km2).
}}

\item{island}{Input value should be either;
\itemize{
\item 0 (= North Island),
\item 1 (= South Island).
}}

\item{introloc}{ASF entry point. Input value should be either;
\itemize{
\item 0 (= Auckland or Picton for the North and South Island, respectively),
\item 1 (= Wellington or Dunedin for the North and South Island, respectively).
}}

\item{control_index}{Control scenarios. Input value should be either;
\itemize{
\item 0 (= no control),
\item 1 (= Baseline control),
\item 2 (= Increase control range),
\item 3 (= Add carcass removal in buffer zone),
\item 4 (= Add passive surveillance in surveillance zone).
}
Default scenario is no control.}

\item{foi}{(optional) A way to calculate force of infection. Input value should be either;
\itemize{
\item 0 (= Han et al (2021)),
\item 1 (= Pepin et al (2022)),
\item 2 (= Thulke et al (2018)).
}
Default method is Han et al (2021).}
}
\value{
This function returns;
\itemize{
\item the number of iterations having ASFV for each cell throughout the simulation period, and
\item the number of habitat cells having ASFV for each week of the simulation period.
}
}
\description{
Simulation of ASF transmission in New Zealand feral pigs
}
\examples{
ASF_simulation(total_iter= 200, ncore= 10, simyear= 5, pop= 1, scale= 0, island= 0, introloc= 0, control_index= 0)
}
