% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coresim.R
\name{coresim}
\alias{coresim}
\title{Core ASF simulation function}
\usage{
coresim(
  input_info,
  input_road,
  input_river,
  input_proxycid,
  input_proxydist,
  input_proxyroad,
  input_proxyriver,
  coreid,
  rep,
  simyear,
  pop,
  scale,
  island,
  introloc,
  control_index,
  foi,
  tmp_map,
  tmp_ts
)
}
\arguments{
\item{input_info}{(internally generated) contains information of cells.}

\item{input_road}{(internally generated) contains the presence of roads between neighbouring cells.}

\item{input_river}{(internally generated) contains the presence of rivers between neighbouring cells.}

\item{input_proxycid}{(internally generated) contains the cell id of proximal cells.}

\item{input_proxydist}{(internally generated) contains the distance to proximal cells.}

\item{input_proxyroad}{(internally generated) contains the presence of roads between proximal cells.}

\item{input_proxyriver}{(internally generated) contains the presence of rivers between proximal cells.}

\item{coreid}{(internally generated) indicates CPU ID for parallel computation.}

\item{rep}{(internally generated) indicates the number of iterations for each core.}

\item{simyear}{Simulation period in year. Input value should be an integer between one and five.}

\item{pop}{Total population size of feral pig. Input value should be either 0 (= small), 1 (= intermediate), or 2 (= large).}

\item{scale}{Average habitat size of feral pigs. Input value should be either 0 (= 4km2) or 1 (= 10km2).}

\item{island}{Input value should be either 0 (= North Island) or 1 (= South Island).}

\item{introloc}{ASF entry point. Input value should be either 0 (= Auckland or Picton for the North and South Island, respectively) or 1 (= Wellington or Dunedin for the North and South Island, respectively).}

\item{control_index}{Control scenarios. Input value should be either 0, 1, 2, 3, or 4. (0: No control. 1: Baseline control. 2: Increase control range. 3: Add carcass removal in buffer zone. 4: Add passive surveillance in surveillance zone).}

\item{foi}{(optional) A way to calculate force of infection. Input value should be either 0, 1, or 2. (0: Han et al (2021). 1: Pepin et al (2022). 2: Thulke et al (2018). Default method is Han et al (2021).}

\item{tmp_map}{(internally generated) is the temporary file to store the spatial result.}

\item{tmp_ts}{(internally generated) is the temporary file to store the temporal result.}
}
\value{
This function returns the number of iterations that each cell having ASFV over the simulation period.
}
\description{
Core ASF simulation function
}
