% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result_map.R
\name{Result_map}
\alias{Result_map}
\title{Draw a figure of ASFV presence on the map}
\usage{
Result_map(res, year, week)
}
\arguments{
\item{res}{An output vector (for single core) / list (for multiple cores) of \code{\link{ASF_simulation}}.}

\item{year}{Year of the chosen timepoint. Input value should be between 1 and 5 which are the minimum and maximum simulation periods in year, respectively.}

\item{week}{Week of the chosen timepoint. Input value should be between 1 and 52.}
}
\description{
This function generates a map of the probability of ASFV presence over the simulated area at a certain timepoint.
}
\details{
In order to generate a map, end-users should designate the timepoint that they want to inspect.
For example, the very first week of the simulation period can be chosen by setting both the year and week as 1.
}
