% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result_ts.R
\name{Result_ts}
\alias{Result_ts}
\title{Draw a time-series of ASFV presence}
\usage{
Result_ts(res)
}
\arguments{
\item{res}{An output vector (for single core) / list (for multiple cores) of \code{\link{ASF_simulation}}.}
}
\value{
A figure of time-series of the total area that ASF virus present. The red line (grey area) indicates the median (interquartile) of the affected area.
\itemize{
\item val Median.
\item low 25th percentile.
\item high 75th percentile.
}
}
\description{
This function generates a time-series of the total area that ASFV present over the simulated area.
}
