% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpercore.R
\name{simpercore}
\alias{simpercore}
\title{Allocating ASF simulations to each core}
\usage{
simpercore(rep, simyear, pop, scale, island, introloc, control_index, foi)
}
\arguments{
\item{rep}{Number of iterations per each core.}

\item{simyear}{Simulation period in year. Input value should be an integer between one and five.}

\item{pop}{Total population size of feral pig. Input value should be either 0 (= small), 1 (= intermediate), or 2 (= large).}

\item{scale}{Average habitat size of feral pigs. Input value should be either 0 (= 4km2) or 1 (= 10km2).}

\item{island}{Input value should be either 0 (= North Island) or 1 (= South Island).}

\item{introloc}{ASF entry point. Input value should be either 0 (= Auckland or Picton for the North and South Island, respectively) or 1 (= Wellington or Dunedin for the North and South Island, respectively).}

\item{control_index}{Control scenarios. Input value should be either 0, 1, 2, 3, or 4. (0: No control. 1: Baseline control. 2: Increase control range. 3: Add carcass removal in buffer zone. 4: Add passive surveillance in surveillance zone).}

\item{foi}{(optional) A way to calculate force of infection. Input value should be either 0, 1, or 2. (0: Han et al (2021). 1: Pepin et al (2022). 2: Thulke et al (2018). Default method is Han et al (2021).}
}
\value{
This function allocates and runs the simulation model to each core.
}
\description{
Allocating ASF simulations to each core
}
